test(agent): move documentation guards where they actually run
The guards landed in tests/eval/, which pytest's unit run never collects and
which CI reaches only through the agent_behavior filter — editing the very
docs they assert would not have run them. Same failure shape as #692's
self-filtered guard. Moved to tests/unit/ alongside
test_secrets_docs_consistency.py; Makefile:83's test-docs recipe repointed.

Mutation-verified both ways:
  ARCHITECTURE.md + 'Next.js'  -> RED (unmarked retired technology) -> revert -> GREEN
  testing-strategy.md coverage -> RED ({'Backend total': 86} != 87) -> revert -> GREEN

Known gap reported, not silently patched: docs/** changes do not trigger
ci-agent (the agent filter covers apps/agent/** and packages/contract/**
only), and the existing precedent has the same hole. Filed separately.
